+++ src/DocumentTemplate/_DocumentTemplate.py
@@ -127,17 +127,17 @@ def join_unicode(rendered):
     """join a list of plain strings into a single plain string,
     a list of unicode strings into a single unicode strings,
     or a list containing a mix into a single unicode string with
-    the plain strings converted from latin-1
+    the plain strings converted from UTF-8
     """
     try:
         return ''.join(rendered)
     except UnicodeError:
         # A mix of unicode string and non-ascii plain strings.
-        # Fix up the list, treating normal strings as latin-1
+        # Fix up the list, treating normal strings as UTF-8
         rendered = list(rendered)
         for i in range(len(rendered)):
             if isinstance(rendered[i], str):
-                rendered[i] = unicode(rendered[i], 'latin-1')
+                rendered[i] = unicode(rendered[i], 'utf-8')
         return u''.join(rendered)